feat(PrintJob): remove LabelPrintJob class; abstract into PrintJob class
